// projects/shared-fp/src/lib/models/shutter-speed.model.ts
export type SlowestShutterSpeedLimit =
  | 'auto-faster'
  | 'auto-fast'
  | 'auto'
  | 'auto-slow'
  | 'auto-slower'
  | '1/8000'
  | '1/6000'
  | '1/5000'
  | '1/4000'
  | '1/3200'
  | '1/2500'
  | '1/2000'
  | '1/1600'
  | '1/1250'
  | '1/1000'
  | '1/800'
  | '1/640'
  | '1/500'
  | '1/400'
  | '1/320'
  | '1/250'
  | '1/200'
  | '1/160'
  | '1/125'
  | '1/100'
  | '1/80'
  | '1/60'
  | '1/50'
  | '1/40'
  | '1/30'
  | '1/25'
  | '1/20'
  | '1/15'
  | '1/13'
  | '1/10'
  | '1/8'
  | '1/6'
  | '1/5'
  | '1/4'
  | '0.3'
  | '0.4'
  | '0.5'
  | '0.6'
  | '0.8'
  | '1';

export type ShutterSpeed =
  | '30 s'
  | '25 s'
  | '20 s'
  | '15 s'
  | '13 s'
  | '10 s'
  | '8 s'
  | '6 s'
  | '5 s'
  | '4 s'
  | '3.2 s'
  | '2.5 s'
  | '2 s'
  | '1.6 s'
  | '1.3 s'
  | '1 s'
  | '0.8 s'
  | '0.6 s'
  | '0.5 s'
  | '0.4 s'
  | '0.3 s'
  | '1/4 s'
  | '1/5 s'
  | '1/6 s'
  | '1/8 s'
  | '1/10 s'
  | '1/13 s'
  | '1/15 s'
  | '1/20 s'
  | '1/25 s'
  | '1/30 s'
  | '1/40 s'
  | '1/50 s'
  | '1/60 s'
  | '1/80 s'
  | '1/100 s'
  | '1/125 s'
  | '1/160 s'
  | '1/200 s'
  | '1/250 s'
  | '1/320 s'
  | '1/400 s'
  | '1/500 s'
  | '1/640 s'
  | '1/800 s'
  | '1/1000 s'
  | '1/1250 s'
  | '1/1600 s'
  | '1/2000 s'
  | '1/2500 s'
  | '1/3200 s'
  | '1/4000 s'
  | '1/5000 s'
  | '1/6000 s'
  | '1/8000 s';